packages/server/requirements.txt
Flask
Flask-SocketIO
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796
python-vlc